requirements.txt
fastapi[all]
sqlmodel
jinja2
python-multipart
passlib[bcrypt]
pycryptodome
python-jose
